% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_headers.R
\name{get_headers_string}
\alias{get_headers_string}
\title{生成需要计算签名的请求头字符串}
\usage{
get_headers_string(headers)
}
\arguments{
\item{headers}{一个named list of character string, 包含所有在请求中的头文件信息}
}
\value{
一个需要包括在签名计算中的请求头字符串, 编码为UTF-8
}
\description{
生成需要计算签名的请求头字符串
}
\details{
需要计算签名的请求头为CanonicalizedOSSHeaders, 所有以x-oss-为前缀的HTTP Header被称为CanonicalizedOSSHeaders
}
\references{
\href{https://help.aliyun.com/document_detail/31951.html#title-wbj-1a2-eyc}{CanonicalizedOSSHeaders构建方式}
}
