% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_resource.R
\name{get_resource_string}
\alias{get_resource_string}
\title{生成需要计算签名的资源字符串}
\usage{
get_resource_string(bucket = "", path = "", query_args = NULL)
}
\arguments{
\item{bucket}{被请求OSS资源的 bucket}

\item{path}{被请求OSS资源的路径(object)}

\item{query_args}{额外的请求参数, 包含自资源的请求参数等}
}
\value{
一个需要被包括在签名计算中的资源字符串
}
\description{
生产需要计算签名的资源字符串
}
\details{
用户发送请求中想访问的OSS目标资源被称为CanonicalizedResource; 需要访问的OSS资源格式为/bucket/path, 如果请求的资源包括子资源（SubResource），则将所有的子资源按照字典序，从小到大排列并以&为分隔符生成子资源字符串
}
\seealso{
\href{https://help.aliyun.com/document_detail/31951.html#title-wnt-ob2-f9f}{构建CanonicalizedResource的方法}
}
