% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.signature-package.R
\docType{package}
\name{oss.signature-package}
\alias{oss.signature-package}
\alias{oss.signature}
\title{Aliyun OSS Request Signatures}
\description{
Generates Aliyun Object Storage Service (OSS) request signatures for RESTful APIs.
}
\details{
This package contains functions mostly intended for developers to use in building API client packages for Aliyun OSS APIs.

The main function of interest is \code{\link{signature_v1_auth}}, which wraps the other internal functions and returns a named list of elements to be used in authenticating an API request using OSS signature.

The package identify credentials by walking through a tree of possible sources of values (described in \code{\link{locate_credentials}}), with optional verbosity, in a manner similar to the Python boto 3 library.

A lower-level function that may be of use to end users is \code{\link{use_credentials}}, which sets the environment variables used by this package based upon values specified in a \samp{.ossutilconfig} file. That function is called by default during package load, if no environment variables are set.
}
\seealso{
\code{\link{signature_v1_auth}}, \code{\link{locate_credentials}}, \code{\link{use_credentials}}
}
\author{
Ruiyang Gan <ganruiyang@metrodata.cn>
}
\keyword{package}
